individual planet stats added

+++ Assets/Scripts/planets/planetStats.cs
@@ -1,11 +1,12 @@
 ﻿using UnityEngine;
 using System.Collections;
+using UnityEditor;
 
 public class planetStats : MonoBehaviour {
 	public float initSpeed = 0;
 	public float orbitRadius = 0;
-	public int population = 0;
-	public int followers = 0;
+	public float population = 0;
+	public float followers = 0;
 	public string name = "planet";
 	public int size = 14;
 	public string biome = "";
@@ -15,7 +16,6 @@ public class planetStats : MonoBehaviour {
 
 	// Use this for initialization
 	void Start () {
-		
 	}
 	
 	// Update is called once per frame
